# Conflicts:
#	geode-core/src/main/java/com/gemstone/gemfire/distributed/internal/InternalDistributedSystem.java

+++ geode-core/src/main/java/com/gemstone/gemfire/internal/cache/tier/sockets/command/Put70.java
@@ -105,7 +105,7 @@ public class Put70 extends Put65 {
     }
     replyMsg.setNumberOfParts(parts);
     replyMsg.setTransactionId(origMsg.getTransactionId());
-    replyMsg.addBytesPart(new byte[]{pr.getMetadataVersion().byteValue(), nwHopType});
+    replyMsg.addBytesPart(new byte[]{pr.getMetadataVersion(), nwHopType});
     replyMsg.addIntPart(flags);
     if (sendOldValue) {
 //      if (logger.fineEnabled()) {
